backend/requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.95.2
greenlet==2.0.2
h11==0.14.0
idna==3.4
passlib==1.7.4
psycopg2-binary==2.9.6
pydantic==1.10.8
PyJWT==2.7.0
python-multipart==0.0.6
sniffio==1.3.0
SQLAlchemy==2.0.15
starlette==0.27.0
typing_extensions==4.6.0
uvicorn==0.22.0
